move unilateral exit into its own file
